presentations: change incorrect presentation methods
The presentation method PRESENT specialized on T is the same as the
default method, so it may be safely removed.

The presentation method PRESENTATION-TYPEP was incorrectly specialized
on T instead of (EQL T). Consider the example:

    (presentation-typep <whatever> t)                ; -> T
    (presentation-typep <whatever> <something-else>) ; -> NIL

The second example returns NIL because both presentation types have
nothing in common.

Both specializations on (EQL T) and (EQL NIL) are moved after the
definition of the presentation type SYMBOL, because that's the
presentation type of these objects.