define-presentation-method: fix the massaged-type the EQL ptype
This is necessary, because otherwise MASSAGED-TYPE for i.e (EQL T)
would be NIL, because T (the type-key-arg value) is not a subclass of
SYMBOL, what is the type name in this situation. Massaged type after
traversing the type lattice would be NIL. The correct massaged type is
always the result of (presentation-type-of object).