presentation-typep: default presentation method signals error
The function PRESENTATION-TYPEP also behaves differently now. When the
presentation type corresponds to the class, first we check with
CL:TYPEP whether the object is of class TYPE. Only after that we test
with the presentation generic functions whether parameters do
match. That better reflects the CLIM specification which states:

"Adding parameters to a presentation type specifier produces a
subtype, which contains some, but not necessarily all, of the objects
that are members of the unparameterized type."

This behavior is also more consistent with CLIM-TOS.