[disabled] presentation-mixin: assert the presentation-object type
This assertion ensures, that the error is signaled when the
presentation is created and the object is not PRESENTATION-TYPEP to
the presentation type.

That change allows us to fail early when the object is presented
instead of erroring at random place (i.e in the presentation
translator body).

An important side effect of this change is also to signal the error,
when we try to present an object with a presentation type which
doesn't have the presentation PRESENTATION-TYPEP implemented. This
saves us from having error later in TEST-PRESENTATION-TRANSLATOR if it
is invoked.

This change is temporarily disabled.