recording: implement copying of coordinate sequences more generally
Instead of recognizing certain arguments names and generating special
code for those, GENERATE-MEDIUM-RECORDING-BODY supports the extended
syntax

  (NAME INITFORM STOREP)

for argument descriptions in which INITFORM specifies a form to
compute the value to be stored in the output record for the argument
and STOREP controls whether a slot corresponding to the argument
should exist in the output record at all.

DEF-GRECORDING also understands this syntax and uses it conditionally
generate slots in the output record class.

The new function COPY-SEQUENCE-INTO-VECTOR is used to copy
sequence-valued arguments into vectors, the idea being that a vector
provides more compact storage when a handful or more elements are
stored.