recording: copy string for DRAW-TEXT-OUTPUT-RECORD in MEDIUM-DRAW-TEXT*
It is not generally correct or safe to store the string passed to
MEDIUM-DRAW-TEXT* in the created DRAW-TEXT-OUTPUT-RECORD instance:

* The string could be changed after the MEDIUM-DRAW-TEXT* call:

    (loop with buffer = (make-string 10)
          do (fill-buffer buffer)
             (draw-text* stream buffer ...))

* The string could have dynamic extent in which case storing it in the
  DRAW-TEXT-OUTPUT-RECORD would lead to out-of-extent accesses:

  (let ((string ...))
    (declare (dynamic-extent string))
    (draw-text* stream string ...))

  This is an actual problem which leads to bad crashes for seemingly
  innocent things like (princ 5 <output-recording-stream>) in recent
  SBCL versions.

The fix implemented in this change ensures that the contents of the
STRING argument is (partially) copied by inserting to SUBSEQ with
appropriate start and end positions. Since the START and END are only
used in the SUBSEQ call, the are no longer stored in the created
output record.