recording: fix DEFRECORD-PREDICATE DRAW-TEXT-OUTPUT-RECORD
The slots START and END have been removed from DRAW-TEXT-OUTPUT-RECORD
in ab85d944 but the methods generated by
(defrecord-predicate draw-text-output-record) still tried to include
them in the comparison.

To fix the issue, DEFRECORD-PREDICATE can now accept things that are
keyword arguments but not slots and can also generate different
comparison code for the slot and the argument cases.