examples: remove invalid method from a stream-test demo
WITH-INPUT-EDITING macro redraws the interactor line, so unless the input buffer
is in transition state this will echo nothing. Example of such behavior is a
backspace character being in the INPUT-BUFFER waiting to be processed only when
STREAM-READ-GESTURE is called. This test behaved fine until now because we don't
have real internal buffers for standard-extended-input-stream objects.