stream-input: separate the input-buffer from the event-queue
We remove an invalid assumption that the INPUT-BUFFER is same thing as the
EVENT-QUEUE for a STANDARD-EXTENDED-INPUT-STREAM. To do that we split initargs
and slots between STANDARD-SHEET-INPUT-MIXIN and the internal input-kernel-mixin
(along with appropriate accessors).

CLIM after this change works as before except that SEIS signals an unbound-slot
condition when we try to read the STREAM-INPUT-BUFFER (not used in the
core). This commit only makes the use of the event queue explicit, it doesn't
change semantics of McCLIM inner workings.