clouseau: wrap all command documentation in safe printing macros
Command documentation can involve printing arbitrary objects which can
lead to printing long sequences, printing circular structure and
calling user-defined, possibly error-signaling code. As a consequence,
wrap all command documentation generation which involves inspected
values in WITH-PRINT-ERROR-HANDLING and WITH-SAFE-AND-TERSE-PRINTING.

This change applies the macros to code that does not strictly need
them in some cases. There should be little harm in doing so.