clouseau: fix COM-{COPY-PLACE-VALUE,SWAP-PLACE-VALUES} commands
The commands and corresponding drag and drop translators were written
before the introduction of the :DESTINATION-TESTER extension to
DEFINE-DRAG-AND-DROP-TRANSLATOR.

This change adds the helper functions TRANSFER-VALUE and
VALID-VALUE-TRANSFER-P for transferring a value (including the unbound
"value") from one place to another and checking the possibility of the
former respectively. The commands use TRANSFER-VALUE, the translator
tests use VALID-VALUE-TRANSFER-P. The latter happens in the
:DESTINATION-TESTER since the object and the destination object are
only available in this tester.