seos: do not assume stream-char-height after a new line
Consider the following snippet:

    (with-text-size (stream :huge)
      (format stream "HEY1!"))
    (fresh-line stream)
    (format stream "HEY2!~%")
    (format stream "HEY3!~%")
    (format stream "HEY4!~%")

When the above code is executed, then FRESH-LINE caused advancing the cursor
vertically by the STREAM-CHAR-HEIGHT of the line with the size :HUGE), and did
not reset the STREAM-CHAR-HEIGHT, so after the line "HEY2!" a vertical
advancement was once again the STREAM-CHAR-HEIGHT line with the size :HUGE.

    (with-text-size (stream :huge)
      (format stream "HEY1!~%"))
    (format stream "HEY2!~%")
    (format stream "HEY3!~%")

The situation is similar, however it is slightly different: the newline is
called when a text style with the size :HUGE is still the context of
drawing. That illustrates, that we should not set STREAM-CHAR-HEIGHT to the
current text-style height after the line break, but rather wait for the first
character written.

The new approach is to reset the STREAM-CHAR-HEIGHT to 0 after each line break
and call MAXF on it only when the text is drawn. After this change the
vertical advancement of the cursor after the new line remains the same (that
is the "old line" height is used), only the initial baseline is set to 0.