basic-port: implement click to focus policy
Port keeps track of the focused sheet and dispatches keyboard events to
it. When the pointer is clicked over a sheet, an event of the type
WINDOW-MANAGER-FOCUS-EVENT is dispatched to it. If sheet wants to take a focus
it should call (SETF PORT-KEYBOARD-INPUT-FOCUS) on itself from the
HANDLE-EVENT method.

TOP-LEVEL-SHEET-MIXIN preserves the keyboard focus from the last time when the
window was focused.

Focused sheet receives keyboard events (if we introduce more input events
which are not tied to the pointer it is very likely that they will be
distributed to the focused sheet too). When the port has no focused sheet then
keyboard events are ignored.