triggering the layout protocol
CHANGE-SPACE-REQUIREMENTS trigger the layout protocol only when
:RESIZE-FRAME keyword is t. Otherwise it calls
NOTE-SPACE-REQUIREMENTS-CHANGED and it is the parent that decides what
to do.

Usually (NOTE-SPACE-REQUIREMENTS-CHANGED parent child)
calls (CHANGE-SPACE-REQUIREMENTS parent), until the message arrives to
the top-level-pane that trigger the layout protocol calling
LAYOUT-FRAME. Exceptions are RESTRAINING-PANE that doesn't do nothing,
and VIEWPORT-PANE that doesn't change its space-requirements but
allocate its space without changing its dimension in order to update
which part of the child is viewed.