default-translator: give a high priority 100
Default translator is the identity translator, that is it returns the same
object it receives. It is applicable only when the presentation type is a
subtype of the context type. Its translation types are T->NIL (wildcard).

Consider that the context type is A, B is a subtype of A and there is a
translator A->B. When a presentation with the type A is selected, then
translator A->B has a higher priority than T->NIL because FROM-TYPE is closer
to the first translator FROM-TYPE (in the supertype inheritance list), so for
the context A a translated object of a type B is returned.

We give the default translator priority 100. It is still possible to force the
translation between two types, but it should not happen by accident. Another
consequence of that is that if the presentation type is a subtype of a context
type, then even if there exists a translator to exact context type, it won't
be used (unless it has a priority higher than 100).