make-menu-item: improve the implementation
- remove a hacky limited evaluator (marked as fixme)

This "evaluator" seems very bogus. It tries to handle a situation, where
someone calls

  (make-menu-item … :text-style '(make-text-style …))

what makes no sense, because a list is not a text style (and I don't know of
any specification entry which would say anything about that). We don't seem to
use that anywhere either.

- call make-instance with all arguments

the function make-menu-item has tried to not apply to initialize-instance
arguments, which were not passed to make-menu-item. That would make sense if:

1. this function weren't internal
2. there were an initialize-instance method which does something clever
3. the implementaiton had anticipated uninitialized slots (it doesn't!)

- initialize all slots in command-item and %menu-item

the implementation in the file does not anticipate uninitialized slots, so we
initialize all of them for classes class-item and %menu-item. The problem is
described by @hershrobe in #955 and he has came up with a fix in #968.