stupid-subtypep: bail on most meta types
According to the specification, FIND-PRESENTATION-TRANSLATORS is not
supposed to consider presentation type parameters but for most pairs
of types involving at least one meta type the subtyping question
cannot be answered without taking the parameters into account. As a
result, STUPID-SUBTYPEP returns the conservative answer (that is true)
for most cases involving at least one meta type.

This behavior is in line with the new behavior of the cache used by
FIND-PRESENTATION-TRANSLATORS.