command parsers: don't try to present unsupplied arguments
This is a partial fix for the issue #955. Instead of presenting the unsupplied
argument with its type we print in italics <unsupplied> instead.

command parsers: bind argument to its value in the unparser

The generated function bound arguments to a result of calling the function
present (or write-string) instead of the actual value. That lead to a
situation, where commands with dependent arguments, like

    (define-my-frame-command (quxbar  :name t :menu t)
        ((arg1 `(clim:sequence integer))
         (arg2 `(member ,@arg1)))
      (format t "integer: ~a ~a~%" arg1 arg2))

Were unparsed improperly, that is (in pseudocode):

    (let ((arg1 (present 42 'integer)
          (arg2 (present 42 '(member)))))
      <body>)

And that lead to an error, because 42 is not of the presentation
type (completion nil). Correct expansion is:

    (let ((arg1 (present 42 'integer)
          (arg2 (present 42 '(member 42)))))
      <body>)

And that leads to a correct expansion. The problem was visible when we were
moving a pointer over already parsed and executed command like this in the
inspector.