accept command: provide input context to click on a command
In the ACCEPT presentation-method for COMMAND:

1. Use WITH-INPUT-CONTEXT for the command's presentation-type so that this
method can handle mouse clicks on objects that translate to (or are presented
as) commands which are acceptable in the current context

2. When the read command is only partial, call the partial command
parser. This may happen for both typed input and when selecting a presentation
with the type command -- it is permissible for displaying incomplete commands.

3. If the input needs to be replaced, then explictly provide the :RESCAN NIL
option to PRESENTATION-REPLACE-INPUT. This has a different effect than not
providing :rescan at all (which seems to default to T).

The reason for 3 is that the command returned by clicking on an object may
contain arguments that print (or present) unreadably.
PRESENTATION-REPLACE-INPUT presents that command into the input buffer and if
:RESCAN NIL is not provided, then ACCEPT-1 will call the ACCEPT presentation
method, invoking the command-parser again. However, this time the input buffer
contains the unreadable string for the argument in the command and the command
argument parser will then signal an error. This was observed in Clouseau.

Co-authored-by: Howard Shrobe <hes@csail.mit.edu>
Co-authored-by: Daniel Kochmański <daniel@turtleware.eu>