filename-completer-get-candidates: remove invalid test
This fixes the problem that requires you to hit Return (enter) twice after
completing a filename.

The previous version had the entire body wrapped inside a (when (null
pathname-type) form that prevented completing if the file-type had been
provided as it would be after hitting the complete gesture (tab).  This test
was incorrect is removed in the fix.

Co-authored-by: Howard Shrobe <hes@csail.mit.edu>