panes: composition: fix box and rack layouts with +FILL+ and spacing
The two BOX-LAYOUT-MIXIN/XICALLY-ALLOCATE-SPACE methods lay out child
panes sequentially, keeping track of the offset in the major
dimension. This offset is increased according to the allotted widths
of children as well as the spacing between children. Before this
change, a +FILL+ client (which does not have a pane and thus does not
need spacing) caused one additional unit of spacing to be added to the
offset, potentially pushing the final "real" client out of the
allocated region.