events: compute the pointer position in sheet coordinates eagerly
Computing the position lazily when the pointer-event-* was called is
incorrect, because the specification says

  Returns the x and y position of the pointer at the time the event occurred.

Previous behavior where coordinates computed "on demand" did not work if the
sheet transformation has changed. For instance:

1. applicable presentation is at Y=4, pointer is at Y=3
2. user scrolls by Y=-1 over Y=3, so the pointer is /now/ at sheet's Y=4
3. scroll gesture activates the command associated with the presentation

We also remove a generic function POINTER-EVENT-POSITION* which was returning
the event coordinates in sheet coordinate system /at the time of calling it/.
This function is not used in McCLIM codebase and is not exported.