stream-panes: set LABEL and SCROLL-BARS slots as obsolete
I think that it is correct remove the LABEL and SCROLL-BARS slots of
CLIM-STREAM-PANE because this information is about CLIM-STREAM-PANE's
ancestors.

The right procedure to set this information it is to use keywords in
the MAKE-CLIM-STREAM-PANE function.

Anyway I didn't remove the slots because other software outside the
McCLIM repository could use them (climacs, ...), instead I set them to
:OBSOLETE and then a warning is signaled if the value is changed
during the initialization of the instance.

Also in McCLIM manual sec. 2.4.1 is stated that :SCROLL-BARS option
for stream panes is obsolete.