drei-gadget-pane: don't put gestures in the input-buffer
Fixes the issue where drei-typed characters appeared in both drei gadget pane
and in the interactor pane. It was caused by that drei-gadget-pane is
standard-input-extended-stream, so after handling the event in the primary
method, an after method appended a gesture to the input-buffer (which is by
default shared by all SEIS in the frame). The appended gesture was read from
the interactor pane and put in the input-editing-stream.