stream-panes: space requirements in MAKE-CLIM-STREAM-PANE
When :scroll-bars is a cons the user space requirements options are
passed to the stream and not to stream container.

You can pass option to the scroller as suggested in Franz manual as:

  (make-clim-stream-application :type 'application-pane :width 500
                                :scroll-bars '(:vertical :width 300))

here 500 is the width of the stream and 300 the width of the scroller.

If :scroll-bars isn't a cons the user space requirement options
belong to the most external container of the stream (scroller-pane,
label-pane or outline-pane) as before this commit.