map-over-command-table-*: provide uniform interface
Previously all command table mapping operators accepted a keyword INHERITED
defaulting to T. Thanks to that it was possible to inhibit inherited objects
mapping. The functions MAP-OVER-COMMAND-TABLE-KEYSTROKES and
MAP-OVER-COMMAND-TABLE-MENU-ITEMS are exceptional in the specification because
the keyword parameter is not specified for them.  As a McCLIM extension,
MAP-OVER-COMMAND-TABLE-KEYSTROKES and MAP-OVER-COMMAND-TABLE-MENU-ITEMS now
also accept the keyword argument.

Also make the function apply-with-command-table-inheritance return NIL,
because its results were not useful.