add-menu-item-to-command-table: improve support for literal sub-menus
The specification mentions only, that a value of the menu item of the type
:menu is is a command table designator. McCLIM extends this model to allow
literal sub-menus as described in the specification of the :menu option of the
macro define-command-table.

Previous version of the code did not work correctly because instead of
creating an anonymous command table with specified items, it put the list
directly in the appropriate slot (so the code working on menu items did not
work).

With this commit defining literal nested sub-menus works as expected. Example
command table definition:

  (define-command-table root-1 :inherit-from nil
    :menu (("Anon X" :menu (("cmd-1" :command `(cmd "Anon cmd 1"))
                            ("cmd-2" :command `(cmd "Anon cmd 2"))))))