(setf frame-command-table): always update the menu bar
When the menu bar exists we always update it when the frame command table
changes. Checking whether the command table has an associated menu is not
correct, because one of the inherited command tables may have a menu too.

Moreover the previous code was wrong, because even when

   (get-menu new-command-table)

would return NIL, that would still demand menu bar update (i.e to remove old
elements).