pointer-grabbing: extend the mechanism
Until now with-grabbed-pointer did not distinguish between grabbing for the
grabbed sheet and for the event owner. The macro WITH-POINTER-GRABBED is
extended to accept a keyword :MULTIPLE-WINDOW - this argument shares the
interpretation with the macro TRACKING-POINTER.

This change is backward compatible from the user perspective, because
:MULTIPLE-WINDOW defaults to NIL, and that means that all events are delivered
to the grabbed sheet.

Also fix syntehsize-boundary-events to return the event sheet if there is none
below the pointer - this situation is the case when we grab the sheet and move
the pointer out of the window.