clx: avoid name collisions for ad-hoc pane classes
Before this change, the names of symbols naming ad-hoc classes were of
the form

  <fm-gensym>-<concrete-pane-class-symbol-name>

which lead to collisions for concrete pane classes named by symbols in
different packages but with identical symbol names. To avoid such
collisions, use symbol names of the form

  <fm-gensym>-<c-p-c-symbol-package-name>:<c-p-c-symbol-name>

Also define the ad-hoc classes by calling MAKE-INSTANCE on the
respective metaclass instead of EVALuating a form.