find-port: change the mechanism for better maintanance
Previously we have stored keyword->class-name mapping in the keyword
plist. This commit preserves the old behavior as a fallback, however the
preferred way to create an instance of the port is to define a method
FIND-PORT-TYPE eql-specialized on the keyword. The function returns two
values: the port class name and the server path parser.