clx: implement sheet and frame icon protocol
The implementation is based on the _NET_WM_ICON property which is
described in the "Application Window Properties" section of the
"Extended Window Manager Hints" specification [1].

The specified mechanism allows putting the pixel data of zero, one or
multiple icons into the property. The X window manager can pick up
these icons and display them to visually augment the X corresponding
window or display the icon instead of the X window, for example when
the X window is iconified.

[1] https://specifications.freedesktop.org/wm-spec/1.5/ar01s05.html#idm45766085139216