data: add default McCLIM logo files
The original logo, mcclim-logo.svg, was created by Pamela Mosiejczuk
<https://gumroad.com/flan_foundry>.  The file in this commit was
retrieved from grabby.space/mcclimlogo.svg and renamed to
mcclim-logo.svg.

The other files have been derived from the above file (and rasterized
in case of the PNG files).

The files are intended to be passed to window managers and desktop
environments as default icons of McCLIM applications that do not have
more specific icons.