frames: fix FRAME-EXIT for three or more nested command loops
This fixes a regression introduced in 96c3158f. Between 96c3158f and
this change, when three or more frames ran in nested command loops and
the outermost frame exited, both inner command loops handled the
FRAME-EXIT condition in the HANDLER-BIND clause. This caused multiple
calls to DISOWN-FRAME of which only the first succeeded.

To reproduce the problem, start demodemo, start three demos, then
close the frame of the demo which has been started first.