mirrors: allow changing transformation of the top-level-sheet-mixin
Previously in %set-mirror-geometry we've ignored top-level-sheet-mixin when it
came to transformations, because we can't always set its exact translation
(for example because of the window manager added decorations).

From now on we do allow moving the top-level-sheet-mixin's mirror and we leave
a comment which notes, that the WM may not honor our request.