clx: realize-mirror: change the clause order for the background ink
We've used to check the sheet first whether it is a
permanent-medium-sheet-output-mixin and after that wheter it is a pane. In the
first case we take the medium-background, while the second has a kludge that
checks whether the background is a color that might be used with clx. Pane may
be a permanent-medium-sheet-output-mixin, so we lift that clause as the first
one.