builtin-commands: use DESCRIBE-PRESENTATION-TYPE as menu label
This change affects the title of the menu frame that is displayed as a
result of the :MENU gesture. Using DESCRIBE-PRESENTATION-TYPE produces
a "human oriented" title and avoids long or even multi-line titles in
case :MENU is invoked on a presentation object with a complex
presentation type such as one containing parameters.