test-presentation-translator: change modifier-state interpretation
When both modifier-state and event are supplied, consequences are
undefined. We commit to an interpretation, that a supplied modifier-state has
a precedence over the event, only only when NIL the event is used.

modifier-state is specified to default to 0 - instead of checking for whether
this argument is supplied in each intermediate function we make the keyword
parameter default to nil, and if neither modifier-state or event is supplied,
then we use the default value.