test-presentation-translator: change the interpretation of arguments
It is unspecified what happens when both modifier-state and event are
supplied. McCLIM exploits this undefined behavior and stipulates, that when
the modifier-state is supplied, then event is ignored.

McCLIM features also an extension that allows testing against a specific
button without consing a new event - this is complementary to the
modifier-state argument. It gains a similar interpretation regarding
modifier-state.

modifier-state defaults to 0, however we always pass it. To avoid juggling
with the modifier-state-p across multiple function along the call chain we
default to NIL and setf the argument to 0 in the test-presentation-translator.