command processing cleanup: ensure that commands are complete
Previously we did that only for the menu items, now we ensure that the read
command is always complete when read by comparing the returned list with a
command canonical form. That way it is consistent with how we treat menu-items
with commands.

Additionally an irregularity is removed where we have created the input
context in which menu-item was handled in the read-frame-command around
method. A translator from the menu-item to the commans is provided, so we may
click the menu-item even when the context accepts only a command. For example:

    (with-input-context ('command :overwrite t)
        (object)
      (something)
      (t object))