execute-frame-command: introduce a command queue
handle-event instead of calling execute-frame-command throws the command with
the appropriate type when the frame is the same as the *application-frame*.
Otherwise the command is enqueued for the next exectution. Handling of the
queue is executed in the :around method defined on the function
read-frame-command.

This solution is to ensure, that the frame is redisplayed after the command
execution. Previously, when we simply called execute-frame-command, the
read-command had no idea that the command has been read and executed, so
read-frame-command did not return and panes were not redisplayed.

This also ensures, that a command is not executed in a context, where we are
waiting for another presentation type (i.e accepting-values, or dragging an
object) and provides some means of synchronization.