execute-frame-command: enqueue command if it is currently read
This commit introduces an internal state of the frame telling, whether a
command is being currently read or not. When the command is being read when
the function execute-frame-command is called, even when it is a synchronous
invocation the execute-command-event is enqueued, so the frame-read-command
may pick it up and return in the appropriate context (ensuring a redisplay).