make-menu-button-from-menu-item: merge two case branches
menu-item-type :command and :function are handler the same way after coercing
the latter to a command.

Check whether the command is enabled by looking directly into the frame's
collection disabled-commands - command-enabled is not called because it checks
whether the command is accessible also in the frame's command table.

Unify behavior for disabled and enabled commands by having the same collection
of armed and disarmed callbacks.