Remove regression caused in 7e6f34296ecc75834e0da34a4c24b4a7393c0237
This commit reverts one change from an old commit. It is not correct for the
stream-read-gesture to trigger redisplay. The commit message from the commit
that introduced this change was not descriptive enough to explain, what
problem this change addresses, so if some issue occurs we'll need to address
it again.