render backend: ensure a correct mirror region
When the window-configuration-event is read, the mirror's image size needs to
change to account for the new region size. This can't be done from the
port-set-mirror-region method, because these calls are inhibited to avoid
infinite loop of setting the mirror region reading it from the port. To
achieve that we add a before method for distribute-event.