standard-presentations: in ACCEPT method for COMPLETION, make type properly
Before this change, the method called MAKE-PRESENTATION-TYPE-SPECIFIER
with the list of presentation type options as the final argument. This
caused the presentation type specified returned as the second value of
the method to retain parameters but omit options compared to the
presentation type specifier passed into the method.

The old overall behavior may or may not have been wrong but it
certainly was inconsistent. The MAKE-PRESENTATION-TYPE-SPECIFIER was
definitely wrong.