Don't execute a disabled command
The spec (26.6.1) say that COMMAND presentation type represent
commands  accessible in COMMAND-TABLE presentation parameter and enabled
in *APPLICATION-FRAME*. The last requirement was not checked before this commit.

The spec (27.4) say that in LOOKUP-KEYSTROKE-COMMAND-ITEM
"only keystrokes that map to an enabled application command will be
matched."

In this way  disabled commands aren't executed.

Fix #1144 issue