clouseau: restrict modification of descendant place when appropriate
In some cases, READ-ONLY-PLACE is not enough: the SYMBOL-NAME of a
symbol is read-only as a place. However, it is also a bad idea to
modify the STRING that is the name of the symbol even if this doesn't
change the value of the symbol name place.

To start addressing this issue, this change adds SUPPORTSP methods
that query ancestor places when determining whether SETF or
REMOVE-VALUE is allowed for a given place. Ancestor places are queried
via (supportsp ANCESTOR 'modify-descendants) which typically traverses
the path to the root place and returns true. However, if an ancestor
along the path returns false, the modification is not allowed. This
way, a symbol name place can to some extent prevent the STRING that is
the name of the symbol from being modified.