panes: unify the interpretation of initargs :align-{x,y}
The specification mentions these initargs as specifiers how child panes are
aligned within the parent pane. Later it mentions that this have the same
semantics as for FORMATTING-CELL.

FORMATTING-CELL wording is sloppy. For example "align-x specifies how the output
in a cell will be aligned relative to _other_ cells in _the same table
column_". This gives the impression that we align cells in its parent while we
align the cell output.

All cells within a row/column have the same size, spacing and alignment --
align-x and align-y are used to position their content.