text recording: improve the initial cursor position
The method tree-recompute-extent specialized the
standard-displayed-output-record used the coordinate of the text bounding
box. We change it to use the initial writing position to avoid a situation,
where incremental redisplay after each iteration detects a different extent of
the output (assuming that the stream region may change to cover all output).

Moreover default margins of the standard-page-layout :top and :left are
absolute (not relative) for a similar reason - when the clim-stream-pane region
grows to account for record before [0,0] we don't want to change the cursor
position by default.