scrolling: fix a regression
Instead of substracting the current scroll bar position we should add it. This
example will assume for simplicity only the translation transformation, but this
reasoning is valid also for other transformations.

  [cur-x, cur-y] is at the position [0, 0] of the viewport

When we call (scroll-extent x y) coordinates are expressed in the sheet
coordianates

  [x, y] is at the position [(- x cur-x), (- y cur-y)] of the viewport

So after transforming [x y] to the viewport coordinates and to scroll to [x y]
we need to move the sheet by -[(+ (- x cur-x) cur-x), (+ (- y cur-y) cur-y)]
because move-sheet takes the bounding rectrangle of the sheet expressed in the
parent coordinates.  Converserly it would be -[x, y] had we assumed only
translations but we are not assuming only translations anymore.