standard-frame-manager: disown all panes when the frame is disowned
Not doing so lead to a situation with the following issue:

1. create a frame
2. run the frame
3. exit the frame
4. run the frame again

The point 4. failed even under the same frame manager (because while disowning
children the layout protocol tried to call compose-space on nil); but most
importantly it was impossible to adopt the frame to a frame manager with a
different type of panes.