medium-draw-polygon*: fix clipping for polygons
Clipping the polygon with a rectangle by computing their intersection may lead
to a region union instance being produced.

During working on this I've noticed that x11 doesn't render correctly polygons
which have a) valid coordinates, b) fall outside of the drawable region; I'm not
going to fix clipping because this backend is going to be deprecated anyway.