grab-pointer: don't use xlib to grab the pointer
Grabbing the pointer at X11 level isn't polite to other applications. The port
already receives all the pointer events on CLIM windows and so a CLIM
application that want track the pointer with :multiple-window t can do it
without grabbing the X11 pointer but only the CLIM pointer.

When a button is pressed X11 automatically grab the pointer so the
button release isn't lost.