generate-panes: frame manager creates  the menu-bar and the pdoc
Previously we've added these panes in the method generated by
define-application-frame, however the menu bar is specified to be handled by the
frame manager (i.e the wm may have a special place for the menu bar).

Pointer-documentation is McCLIM extension which is meant to behave similarily,
so we treat it similar to menu-bar option.