recording: map-over-output-records: consider all children
Before this commit map-over-output-records (and output-record-children) did not
account for children that were not included in the bounding rectangle of the
record:

- empty compound output records
- some children during recompute-extent-for-changed-child

That lead to incorrect mapping and in the latter case incorrect bounding
rectangle recomputation. Here we use the rectangle without bounds and pass it to
the spatial-trees library.