ports: introduce a function port-set-mirror-geometry
For now the default method calls to port-set-mirror-region and
port-set-mirror-transformation however it will be changed soon.

The function takes the region which is the region in the parent mirror
coordinates that must be displayed on this mirror. The function returns mirror
rectangle after it is rounded (according to the mirror approximation rules).