Backend/Emacs: Send presentation IDs and bounding boxes to Emacs
This is a first cut at informing Emacs about the presentations that
exist in the image. Each presentation is given a unique ID number that
is recorded on the Lisp side. Each presentation is also given a
geometric area (for simplicity the bounding box but Emacs supports
circles and polygons too.)

Emacs then attaches these presentation properties as :map attributes
on the image i.e. "hot" areas that can respond to mouse events.

It doesn't seem to be quite working yet on my examples. I suspect
there is a bad interaction with the viewBox SVG attribute where the
graphics are clipped and end up in a different coordinate space to the
:map hot areas.

Have to debug...