define-application-frame: define a method to reinitialize the frame
The (reinitialize-instance :after) method updates constructors of layouts and
panes in the frame instance (and available layouts), and if the frame has
already created panes it tries to reinitialize each of them. After that the
layout is re-applied to the frame.

The call to the method reinitialize-instance specialized on the standard-frame
is thread-safe.

Co-authored-by: Jan Moringen <jmoringe@techfak.uni-bielefeld.de>