clx: xrender: uniform access to xrender picture and gcontext
The picture object is lazily created for each mirror in the drawable's
plist. Getting source and target pictures is done by means of calling
medium-{target,source}-picture. This is a preliminary step towards further
improvements of the xrender rendering.

To use legacy drawing routines (without breaking compatibility with the "legacy"
medium) a new function clx-drawable-gcontext is created. It works similar to
clx-drawable-picture.

Also don't assume that the first picture and the first gcontext on the drawable
plist is "ours" - instead find them with getf.