clx: xrender: introduce a new ink sourcing model
A macro with-render-context macro is responsible for binding the source and the
target ink as well as for setting the target's clip mask. All allocated
resources are released after use. Currently we clip by means of a gcontext.

For now only filled rectangles and text use this macro. Non-uniform designs are
implemented. Fixes #1132.