core: medium-draw-circle*: remove unnecessary protocol
This protocol is redundant with medium-draw-ellipse*. Moreover it is forced to
fallback to drawing ellipses for some medium transformations. On top of all that
draw-circle* already calls medium-draw-ellipse* so this is not used in pracitce.

This change is backward incompatible, because medium-draw-circle* was exported
from the extension and the backend packages.