clx: factor out drawing routines from the medium protocols
This step is necessary to allow using custom gcontext with legacy drawing
routines. This will be used for:

- setting the clipping region
- computing masks in the new ink sourcing model (xrender)

clx-draw-* routines are similar to medium-draw-* however they accept explicit
arguments for the drawable, the gcontext and the transformation. They are
responsible for clipping coordinates to valid x11 values.

Moreover it is assumed that after the transformation the nature of a drawn
object won't change (i.e clx-draw-rectangle will always call xlib:draw-rectangle
even when the transformation is not rectilinear - it is medium-draw-rectangle*
responsiblity to handle this situation.