update frame-state from windowing system events
The frame-state can change due to user interaction with windowing
system. This commit introduce two new events:
 - window-manager-iconify-event
 - window-manager-deiconify-event

The port synthesize these events when necessary and the
top-level-sheet-pane handle them calling shrink-frame or
enable-frame as necessary.

In this way the frame-state of an application frame is updated to the
real state of the frame.

It is also added a comment to note that shrink-frame doesn't work on
all window-managers. The problem was present also before this commit.