core: find-presentation-translator: look in inherited tables
This function is specified to return two values: the presentation translator
and the source table. The previous version returned only one value and did not
look in inherited command tables.

Also since both find-presentation-translator and find-menu-item possibly look
in inherited command tables they should signal command-not-accessible if
errorp is true.