core: command table menu: rewrite destructive operators
- ADD-COMMAND-TO-COMMAND-TABLE when errorp is t first checks for menu names
and keystrokes in the command table (when these options are specified) and
errors before adding the command if they are present. When errorp is nil then
potential entries are first removed to avoid duplicated entries

- REMOVE-COMMAND-FROM-COMMAND-TABLE mistakenly took the command-item-name as a
key to the hash table which key is a command-line-name - fixed

- ADD-MENU-ITEM-TO-COMMAND-TABLE did not check whether the keystroke is
present as an accelerator in other entries - now it does and either errors or
removes it depending on the errorp value

- REMOVE-MENU-ITEM-FROM-COMMAND-TABLE did not remove the menu item if it was
associated with a keystroke. This situation could happen only if both were
defined with the same operator, so removing it makes more sense, consider

  (add-menu-item-to-command-table "Open file" :command 'com-open-file
                                  :keystroke :foo)
  (remove-menu-item-from-command-table "Open file")

when the programmer follows remove after add-menu-item they would suspect that
the access to this command is removed.

- fix numerous calls to find-keystroke-item that tested with equal to call
ensure-physical-gesture on the gesture - otherwise the test fails. This fixed
among other things the function REMOVE-KEYSTROKE-FROM-COMMAND-TABLE.