core: medium-draw-ellipse*: ensure that angles are normalized
The problem was that when start angle and end angle were passed to the
function medium-draw-ellipse* they could have been not normalized - for
example [-2pi, 0] denotes a full circle. Normalizing by (mod angle (* 2 pi))
doesn't work exactly right, because for a full angle we'll have [0, 0] after
the normalization. To ensure both start-angle to be <0,2pi) and end-angle to
be <0,2pi> we introduce a function normalize-angle* that first checks the
distance between both boundaries and then returns two values that respect the
angle width. Fixes #1218.