ttf fonts: map default text styles with cl-dejavu fonts
This saves us from a frequent issue that the system fonts can't be
found. cl-dejavu is available in quicklisp and contains prepackaged
dejavu fonts.

Changes:
- the variable *families/faces* always contains absolute pathnames
- fonts are autoconfigured in eval-when (not in the asd file)

The variable *truetype-font-path* is preserved because it is used to
list the "system" fonts (and i.e to customize this path to load the
"user" fonts).

The fc-match implementation is preserved because it is used by the
device-text-style mapping to files.

Fixes #719.