accept: expression: remove defunct openmcl extension
The presentation type expression had an option 'subform-read' that was
supposed to allow reading subforms. Despite numerous tries I was unable
to make it work under openmcl (i.e the behavior was no different than
without this extension), and code to achieve that was gross, so I'm
removing this implementation-specific extension for maintanace reasons.