accept: expression: don't hardcode returning the expression
Presentation methods accept specialized on the expression returned the
ptype based on the following form:

    (setq ptype (presentation-type-of object))
    (unless (presentation-subtypep ptype 'expression)
      (setq ptype 'expression))

This is not correct because the requested type may be a subtype of the
expression. In that case we should check whether the object's type is a
subtype of that type and otherwisre return the requested type.

    (setq ptype (presentation-type-of object))
    (unless (presentation-subtypep ptype type)
      (setq ptype type))

It is the accept method responsibility of accepting only objects that
are valid in this input context, so we don't need to check

    (assert (presentation-typep object type))

Fixes #452.