core: parsers: change the function keyword-arg-name-form-symbol
This function was used by the keyword acceptor and the keyword unparser. It
converted :key-like-this to :Key Like This (like command-name-form-symbol).
This doesn't work well because if there were multiple keyword arguments with
the same prefix with a dash parsing failed after the first space. I.e

    (define-test-frame-command (command-4 :name "Print that")
        (&key
         (key-with-dash-1 'string)
         (key-with-dash-2 'string))
      (print (cons key-with-dash-1 key-with-dash-2)))

After typing in the parse the command

     Command: Print that (keywords) :key<space>

The parsing error was published because delimiter tried completion and upon
failure the input could not be matched. Moreover among two the second one is
more readable to the human:

     Command: Print that (keywords) :key with dash 1 42
     Command: Print that (keywords) :key-with-dash-1 42

On top of that the spec doesn't seem to require us to convert keywords like
command names (in fact clim-tos doesn't do that either). This commit reads
keywords with dashes like they should be read.